requirements.txt
jinja2
python-dotenv
requests
selenium